requirements.txt
streamlit
langchain
langchain-groq
langchain-community
faiss-cpu
sentence-transformers
PyPDF2
python-dotenv